+++ src/bumper-engine-decorator.js
@@ -1,6 +1,7 @@
 // @flow
 import {FakeEvent, EventType} from '@playkit-js/playkit-js';
-import {Bumper} from './bumper';
+import {Bumper, BumperBreakType} from './bumper';
+import {BumperState} from './bumper-state';
 
 /**
  * Engine decorator for bumper plugin.
@@ -17,7 +18,7 @@ class BumperEngineDecorator implements IEngineDecorator {
   }
 
   get active(): boolean {
-    return this._plugin.playOnMainVideoTag() && this._plugin.isAdPlaying();
+    return this._plugin.playOnMainVideoTag() && (this._plugin.isAdPlaying() || this._plugin.state === BumperState.LOADING);
   }
 
   dispatchEvent(event: FakeEvent): boolean {
@@ -44,7 +45,7 @@ class BumperEngineDecorator implements IEngineDecorator {
    * @memberof BumperEngineDecorator
    */
   get ended(): boolean {
-    return this._plugin.adBreakPosition === -1;
+    return this._plugin.adBreakPosition === BumperBreakType.POSTROLL;
   }
   /**
    * Get the current time in seconds.